Enum rumblebars::SafeWriting

pub enum SafeWriting<'a> {
    Safe(&'a mut SafeWriter + 'a),
    Unsafe(&'a mut Write + 'a),
}

Enum that holds either a writer or a wrapped writer, this is needed to opt-out filtering easily, so rumblebar uses SafeWriting to make use of into_safe() mostly transparent and stateless, while enforcing safe writing through the api.

Typically, the wrapped writer will filter unsafe string/characters.

Examples

Simple html escaping and unsafe support :

use std::io::Write;
use rumblebars::SafeWriting;
let mut buf = Vec::new();
 
SafeWriting::with_html_safe_writer(&mut buf, &|out| {
  out.write_all("<>".as_bytes());
  out.into_unsafe().write_all("<>".as_bytes()) 
});
 
assert_eq!(String::from_utf8(buf).unwrap(), "&lt;&gt;<>")

Double escaping will work as expected :

use std::io::Write;
use rumblebars::HTMLSafeWriter;
use rumblebars::SafeWriting;
let mut buf = Vec::new();
{
  let mut esc1 = HTMLSafeWriter::new(&mut buf);
  SafeWriting::with_html_safe_writer(&mut esc1, &|out| {
    out.write_all("<>".as_bytes())
  });
}
 
assert_eq!(String::from_utf8(buf).unwrap(), "&amp;lt;&amp;gt;")

Variants

Safe

Unsafe

Methods

impl<'a> SafeWriting<'a>

fn into_unsafe(&mut self) -> SafeWriting

disclose the unsafe writer without touching self

fn with_html_safe_writer(out: &mut Write, safe: &Fn(&mut SafeWriting) -> HBEvalResult) -> HBEvalResult

quick way to get a SafeWriter that escapes html