Function rumblebars::parse

pub fn parse(template: &str) -> ParseResult

Parses a handlebars template.

Failures

Handlebars syntax does not have much corner cases, so hopefully you won't get much cryptic cases. The most painfull ones are unmatched blocks opening/closing, this class of errors comes with a short description.

Examples

Provided that you have continous integration with tests on your code, unwraping is safe for static inline templates.

::rumblebars::parse("{{hello}}").unwrap();

Otherwise check for errors.

assert!(::rumblebars::parse("{{#hello}}{{/end}}").is_err());