Struct rumblebars::HelperOptions

pub struct HelperOptions<'a> {
    pub context: &'a HBData + 'a,
    pub condition: bool,
    // some fields omitted
}

Provides rendering and lookup features to registered helpers

Fields

context

current expansion data context

condition

if a parameter was given to helper, holds the result of its truthy/falsy evaluation

Methods

impl<'a> HelperOptions<'a>

fn option_by_name(&self, name: &String) -> Option<&'a HBData + 'a>

data passed as options (eg {{helper option="one" option=name.val}})

fn lookup(&self, key: &HBData) -> Option<&'a HBData + 'a>

fetch data at given path

Examples

use rumblebars::EvalContext;

let mut context = EvalContext::new();

context.register_helper("fetch_hello".to_string(), Box::new(
  |params, options, out, hb_context| {
    // if context is {"hello": {"content": "hi"}}, returns Some("hi")
    options.lookup(&"hello.content");
    Ok(())
}));

fn lookup_with_context(&self, key: &HBData, context: &HBData) -> Option<&'a HBData + 'a>

same as lookup, but with custom context instead of current one (this)

fn render_fn(&self, out: &mut SafeWriting) -> HBEvalResult

for a helper with a block, such as {{if}}{{/if}}, renders block content

fn render_fn_with_context(&self, data: &HBData, out: &mut SafeWriting) -> HBEvalResult

same as render_fn, but with custom context

fn inverse(&self, out: &mut SafeWriting) -> HBEvalResult

for a helpers with blocks, such as {{if}}{{else}}{{/if}}, renders the {{else}} block content

fn inverse_with_context(&self, data: &'a HBData, out: &mut SafeWriting) -> HBEvalResult

same as inverse, but with custom context

fn render_fn_with_context_and_globals(&self, data: &HBData, out: &mut SafeWriting, globals: &HashMap<&str, &HBData>) -> HBEvalResult

allow block rendering with custom context and custom globals (data available with @key)

Examples

use rumblebars::EvalContext;
use rumblebars::HBData;
use ::std::default::Default;

let mut context: EvalContext = Default::default();
let data = "Some Data".to_string();
context.register_helper("fetch_hello".to_string(), Box::new(
  move |params, options, out, hb_context| {
    let mut globals = ::std::collections::HashMap::<&str, &HBData>::new();
    globals.insert("@some_key", &data);
    // if block contains {{@some_key}}, it'll expand to 'Some Data'
    options.render_fn_with_context_and_globals(options.context, out, &globals)
}));

fn render_fn_with_globals(&self, out: &mut SafeWriting, globals: &HashMap<&str, &HBData>) -> HBEvalResult