Struct rumblebars::Template

pub struct Template {
    pub entries: Entries,
}

Internal representation of a handlebars template used for later expansion.

Provides API shortcuts, parsing and expanding are internally processed by ::rumblebars::parse() and ::rumblebars::eval().

(see crate docs)

Fields

entries

Methods

impl Template

fn new(template: &str) -> ParseResult

fn eval_to_string(&self, data: &HBData) -> Option<String>

fn eval(&self, data: &HBData, out: &mut Write, eval_context: &EvalContext) -> HBEvalResult